Move template code
