+++ spec/GHC/Plugin/OllamaHoles/Data/Flags/Types/Gen.hs
@@ -27,7 +27,7 @@ import GHC.Plugin.OllamaHoles.Backend
 import GHC.Plugin.OllamaHoles.Data.Flags.Types
 import GHC.Plugin.OllamaHoles.Data.Trigger.Types
 import GHC.Plugin.OllamaHoles.Logger
-import GHC.Plugin.OllamaHoles.Template
+import GHC.Plugin.OllamaHoles.Data.Template
 
 
 genModelNameText :: QC.Gen Text
