+++ src/GHC/Plugin/OllamaHoles/Data/Flags/Error.hs
@@ -6,7 +6,7 @@ import GHC.Driver.Plugins (CommandLineOption)
 import Data.Text (Text)
 
 import GHC.Plugin.OllamaHoles.Data.Trigger.Error (TriggerPolicyError)
-import GHC.Plugin.OllamaHoles.Template (TemplateError)
+import GHC.Plugin.OllamaHoles.Data.Template.Error (TemplateError)
 
 import GHC.Plugin.OllamaHoles.Data.Flags.Types (FlagName)
