+++ src/GHC/Plugin/OllamaHoles/Data/Flags/Parse.hs
@@ -17,7 +17,7 @@ import GHC.Plugin.OllamaHoles.Data.Trigger.Parse (parseTriggerPolicy)
 
 import GHC.Plugin.OllamaHoles.Data.Flags.Types
 import GHC.Plugin.OllamaHoles.Data.Flags.Error (FlagError(..))
-import GHC.Plugin.OllamaHoles.Template (parseTemplateName)
+import GHC.Plugin.OllamaHoles.Data.Template.Parse (parseTemplateName)
