+++ src/GHC/Plugin/OllamaHoles/Data/Template.hs
@@ -1,0 +1,13 @@
+module GHC.Plugin.OllamaHoles.Data.Template
+  ( module GHC.Plugin.OllamaHoles.Data.Template.Types
+  , module GHC.Plugin.OllamaHoles.Data.Template.Error
+  , module GHC.Plugin.OllamaHoles.Data.Template.Parse
+  , module GHC.Plugin.OllamaHoles.Data.Template.Load
+  , module GHC.Plugin.OllamaHoles.Data.Template.Expand
+  ) where
+
+import GHC.Plugin.OllamaHoles.Data.Template.Types
+import GHC.Plugin.OllamaHoles.Data.Template.Error
+import GHC.Plugin.OllamaHoles.Data.Template.Parse
+import GHC.Plugin.OllamaHoles.Data.Template.Load
+import GHC.Plugin.OllamaHoles.Data.Template.Expand
