+++ src/GHC/Plugin/OllamaHoles/Data/Template/Error.hs
@@ -1,0 +1,19 @@
+module GHC.Plugin.OllamaHoles.Data.Template.Error where
+
+import Data.Text (Text)
+
+import GHC.Plugin.OllamaHoles.Data.Template.Types
+
+
+
+data TemplateError
+    = TemplateFileNotFound FilePath
+    | UnknownTemplateName FilePath Text
+    | UnknownPlaceholders [Placeholder]
+    | MalformedTemplate Line Col TemplateParseError
+    | InvalidTemplateName Text
+    deriving (Eq, Show)
+
+data TemplateParseError
+    = MalformedPlaceholder Text
+    deriving (Eq, Ord, Show)
